+++ test/couch-db-integration.test.ts
@@ -20,6 +20,7 @@ import request from 'request-promise-native';
 
 import * as humansCollection from './helper/humans-collection';
 import * as schemaObjects from './helper/schema-objects';
+import { setDefaultStorage } from './unit/config';
 
 describe('couchdb-db-integration.test.js', () => {
     const COUCHDB_URL = 'http://127.0.0.1:5984/';