+++ test/typings.test.ts
@@ -228,22 +228,12 @@ describe('typings.test.js', function () {
                             ignoreDuplicate: false
                         };
                         const myDb: RxDatabase = await createRxDatabase(databaseCreator);
-
-
                         const minimalHuman: RxJsonSchema<DefaultDocType> = ${JSON.stringify(schemas.humanMinimal)};
-
-
                         const myCollections = await myDb.addCollections({
                             humans: {
                                 schema: minimalHuman,
                             }
                         });
-
-
-
-
-
-
                         await myDb.destroy();
                     })();
                 `;
@@ -397,44 +387,6 @@ describe('typings.test.js', function () {
                             foo: 'bar'
                         });
                         const docs = await myCollection.storageInstance.internals.pouch.allDocs();
-                    })();
-                `;
-                await transpileCode(code);
-            });
-            it('access PouchSyncHandler', async () => {
-                const code = codeBase + `
-                    (async() => {
-                        const myDb: RxDatabase = await createRxDatabase({
-                            name: 'mydb',
-                            storage: getRxStoragePouch('memory'),
-                            multiInstance: false,
-                            ignoreDuplicate: false,
-                            options: {
-                                foo1: 'bar1'
-                            }
-                        });
-                        const mySchema: RxJsonSchema<any> = ${JSON.stringify(schemas.human)};
-                        type docType = {
-                                foo: string
-                        };
-                        const cols = await myDb.addCollections({
-                            humans: {
-                                schema: mySchema
-                            }
-                        });
-                        const myCollection: RxCollection<docType> = cols.humans;
-                        const replicationState = myCollection.syncCouchDB({
-                            remote: 'http://localhost:9090/'
-                        });
-                        const syncHandler = replicationState._pouchEventEmitterObject;
-                        if(!syncHandler) {
-                            process.exit();
-                        }
-                        syncHandler.on('paused', (anything: any) => {
-
-                        });
-                        console.log('.4');
-                        process.exit();
                     })();
                 `;
                 await transpileCode(code);