+++ test/unit/cross-instance.test.ts
@@ -99,7 +99,6 @@ config.parallel('cross-instance.test.js', () => {
 
                 await AsyncTestUtil.waitUntil(async () => {
                     if (emitted.length > 1) {
-                        console.dir(emitted);
                         throw new Error('got too many events ' + emitted.length);
                     }
                     return emitted.length === 1;
@@ -130,6 +129,9 @@ config.parallel('cross-instance.test.js', () => {
             c2.database.destroy();
         });
         it('get no changes via pouchdb on different dbs', async () => {
+            if (config.storage.name !== 'pouchdb') {
+                return;
+            }
             const c1 = await humansCollection.create(0);
             const c2 = await humansCollection.create(0);
             let got;